更改 ConversionRates 属性类型为 Decimal
在 `ExChangeRateResult.cs` 文件中的 `ExChangeRateResult` 类中，`ConversionRates` 属性的类型从 `IDictionary<String, Double>?` 更改为 `IDictionary<String, Decimal>?`。

+++ Pek.ExChangeRate/ExChangeRateResult.cs
@@ -57,5 +57,5 @@ public class ExChangeRateResult
     /// 键为货币代码，值为对应的转换率。
     /// </summary>
     [JsonPropertyName("conversion_rates")]
-    public IDictionary<String, Double>? ConversionRates { get; set; }
+    public IDictionary<String, Decimal>? ConversionRates { get; set; }
 }
